executor: disable TestAnalyzeFastSample parallel
